Introducing FxAWebViewModel (#6733)
* removing logic from controller

* removing webview within viewModel

* inject LeanPlumClient via init

* cleanup

* cleanup

* applying pr changes

* applying second waves of changes
